Projectiles now do damage. Still needs a large overhaul, much of this code should be temporary.
- Added the Boss as a Prefab
- DMGDone and Health is now int instead of float
- Removed Damagable, it pretty much does the same thing as Health
- Added an Owner variable to DMG, this is set when projectiles/hitboxes are made
- Made Projectile script assign Owner of summoned projectile
- Instead of DMG checking for tags, it checks to see is colliding GameObject has a health script, then checks to make sure it is not the owner of this projectile
- Tweaked some Character prefab stuff that just isn't commiting

+++ Assets/Scripts/Health.cs
@@ -24,7 +24,7 @@ public class Health : MonoBehaviour {
 
 
 	// Apply damage to this object.
-	public void ApplyDMG(float DMG) {
+	public void ApplyDMG(int DMG) {
 		HealthPoints -= DMG;
 	}
